ai-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==10.4.0
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4
